+++ patient-generated-data/src/main/java/gov/va/api/health/patientgenerateddata/observation/ObservationController.java
@@ -1,11 +1,12 @@
-package gov.va.api.health.patientgenerateddata;
+package gov.va.api.health.patientgenerateddata.observation;
 
 import static com.google.common.base.Preconditions.checkState;
 import static gov.va.api.health.patientgenerateddata.SerializationUtils.deserializedPayload;
 
 import gov.va.api.health.autoconfig.configuration.JacksonConfig;
 import gov.va.api.health.autoconfig.logging.Loggable;
-import gov.va.api.health.r4.api.resources.QuestionnaireResponse;
+import gov.va.api.health.patientgenerateddata.Exceptions;
+import gov.va.api.health.r4.api.resources.Observation;
 import java.util.Optional;
 import javax.validation.Valid;
 import lombok.AllArgsConstructor;
@@ -26,12 +27,12 @@ import org.springframework.web.bind.annotation.RestController;
 @Validated
 @RestController
 @RequestMapping(
-    value = "/r4/QuestionnaireResponse",
+    value = "/r4/Observation",
     produces = {"application/json", "application/fhir+json"})
 @AllArgsConstructor(onConstructor_ = @Autowired)
-public class QuestionnaireResponseController {
+public class ObservationController {
 
-  private final QuestionnaireResponseRepository repository;
+  private final ObservationRepository repository;
 
   @InitBinder
   void initDirectFieldAccess(DataBinder dataBinder) {
@@ -39,28 +40,27 @@ public class QuestionnaireResponseController {
   }
 
   @GetMapping(value = "/{id}")
-  QuestionnaireResponse read(@PathVariable("id") String id) {
-    Optional<QuestionnaireResponseEntity> maybeEntity = repository.findById(id);
-    QuestionnaireResponseEntity entity = maybeEntity.orElseThrow(() -> new Exceptions.NotFound(id));
-    return deserializedPayload(id, entity.payload(), QuestionnaireResponse.class);
+  Observation read(@PathVariable("id") String id) {
+    Optional<ObservationEntity> maybeEntity = repository.findById(id);
+    ObservationEntity entity = maybeEntity.orElseThrow(() -> new Exceptions.NotFound(id));
+    return deserializedPayload(id, entity.payload(), Observation.class);
   }
 
   @SneakyThrows
   @PutMapping(value = "/{id}")
   @Loggable(arguments = false)
   ResponseEntity<Void> update(
-      @PathVariable("id") String id,
-      @Valid @RequestBody QuestionnaireResponse questionnaireResponse) {
-    String payload = JacksonConfig.createMapper().writeValueAsString(questionnaireResponse);
-    checkState(id.equals(questionnaireResponse.id()), "%s != %s", id, questionnaireResponse.id());
-    Optional<QuestionnaireResponseEntity> maybeEntity = repository.findById(id);
+      @PathVariable("id") String id, @Valid @RequestBody Observation observation) {
+    String payload = JacksonConfig.createMapper().writeValueAsString(observation);
+    checkState(id.equals(observation.id()), "%s != %s", id, observation.id());
+    Optional<ObservationEntity> maybeEntity = repository.findById(id);
     if (maybeEntity.isPresent()) {
-      QuestionnaireResponseEntity entity = maybeEntity.get();
+      ObservationEntity entity = maybeEntity.get();
       entity.payload(payload);
       repository.save(entity);
       return ResponseEntity.ok().build();
     }
-    repository.save(QuestionnaireResponseEntity.builder().id(id).payload(payload).build());
+    repository.save(ObservationEntity.builder().id(id).payload(payload).build());
     return ResponseEntity.status(HttpStatus.CREATED).build();
   }
 }
