+++ patient-generated-data/src/main/java/gov/va/api/health/patientgenerateddata/observation/ObservationRepository.java
@@ -1,4 +1,4 @@
-package gov.va.api.health.patientgenerateddata;
+package gov.va.api.health.patientgenerateddata.observation;
 
 import gov.va.api.health.autoconfig.logging.Loggable;
 import org.springframework.data.repository.CrudRepository;
@@ -7,5 +7,4 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Loggable
 @Transactional(isolation = Isolation.READ_UNCOMMITTED)
-public interface QuestionnaireResponseRepository
-    extends CrudRepository<QuestionnaireResponseEntity, String> {}
+public interface ObservationRepository extends CrudRepository<ObservationEntity, String> {}
