+++ patient-generated-data/src/main/java/gov/va/api/health/patientgenerateddata/questionnaire/QuestionnaireController.java
@@ -1,10 +1,11 @@
-package gov.va.api.health.patientgenerateddata;
+package gov.va.api.health.patientgenerateddata.questionnaire;
 
 import static com.google.common.base.Preconditions.checkState;
 import static gov.va.api.health.patientgenerateddata.SerializationUtils.deserializedPayload;
 
 import gov.va.api.health.autoconfig.configuration.JacksonConfig;
 import gov.va.api.health.autoconfig.logging.Loggable;
+import gov.va.api.health.patientgenerateddata.Exceptions;
 import gov.va.api.health.r4.api.resources.Questionnaire;
 import java.util.Optional;
 import javax.validation.Valid;
