+++ patient-generated-data/src/main/java/gov/va/api/health/patientgenerateddata/questionnaire/QuestionnaireRepository.java
@@ -1,4 +1,4 @@
-package gov.va.api.health.patientgenerateddata;
+package gov.va.api.health.patientgenerateddata.questionnaire;
 
 import gov.va.api.health.autoconfig.logging.Loggable;
 import org.springframework.data.repository.CrudRepository;
@@ -7,4 +7,4 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Loggable
 @Transactional(isolation = Isolation.READ_UNCOMMITTED)
-public interface ObservationRepository extends CrudRepository<ObservationEntity, String> {}
+public interface QuestionnaireRepository extends CrudRepository<QuestionnaireEntity, String> {}
